requirements.txt
aiofiles==0.4.0
asgiref==3.2.3
Click==7.0
click-default-group==1.2.2
datasette==0.37.1
Django==3.0.2
djangorestframework==3.11.0
h11==0.9.0
httptools==0.1.1
hupper==1.10.2
importlib-metadata==1.5.0
janus==0.4.0
Jinja2==2.10.3
joblib==0.15.1
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
Pillow==7.1.2
Pint==0.11
pluggy==0.13.1
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.23.1
scipy==1.4.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
zipp==3.1.0
